Fix compilation warning in i386
platform/drm/cog-platform-drm.c:492:13: warning:
format ‘%ld’ expects argument of type ‘long int’, but argument 4 has type ‘unsigned int’

(cherry picked from commit 48fdd6e8a783e0d396d55e646a17c2cdfb62da4a)

+++ platform/drm/cog-platform-drm.c
@@ -490,7 +490,7 @@ init_drm(void)
         return FALSE;
 
     g_debug("init_drm: using mode [%ld] '%s' @ %dHz",
-            (drm_data.mode - drm_data.connector.obj->modes) / sizeof(drmModeModeInfo *), drm_data.mode->name,
+            (long)((drm_data.mode - drm_data.connector.obj->modes) / sizeof(drmModeModeInfo *)), drm_data.mode->name,
             drm_data.mode->vrefresh);
 
     for (int i = 0; i < drm_data.base_resources->count_encoders; ++i) {
